refactor(frontend): decouple productivity screen logic and ui
- Extracted complex state and filtering logic into `useProductivityViewModel`
- Split `ProductivityHeader` and `ProductivityEmptyState` into separate component files
- Resolved SRP violations by enforcing one logical refactor per file
- Maintained exact behavioral integrity and i18n support
